fix: restructure talis cdk stack and aspects
