+++ lib/aspects/resource_prefixer/cfn_resource_prefixer.ts
