+++ lib/aspects/resource_prefixer/index.ts
