+++ lib/aspects/resource_prefixer/prefixers/dynamodb_cfn_table_prefixer.ts
