+++ lib/aspects/resource_prefixer/resource_prefixer.ts
