+++ lib/index.ts
@@ -1,4 +1,5 @@
 export * from "./authenticated-api";
 export * from "./cdn-site-hosting";
 export * from "./lambda-worker";
-export * from "./common";
+export * from "./talis-cdk-stack";
+export * from "./aspects";
