+++ lib/talis-cdk-stack/index.ts
@@ -2,4 +2,3 @@ export * from "./talis-region";
 export * from "./talis-deployment-environment";
 export * from "./talis-cdk-stack-props";
 export * from "./talis-cdk-stack";
-export * from "./aspects";
