+++ lib/talis-cdk-stack/talis-cdk-stack-props.ts
